+++ src/common/transaction.cpp
@@ -4,7 +4,7 @@
 namespace solstice::matching
 {
 
-Transaction::Transaction(OrderPtr bid, OrderPtr ask, double price, double qnty)
+Transaction::Transaction(OrderPtr bid, OrderPtr ask, double price, int qnty)
     : d_uid(Random::getRandomUid()),
       d_timeExecuted(timeNow()),
       d_bidUid(bid->uid()),
@@ -20,7 +20,7 @@ int Transaction::bidUid() const { return d_bidUid; }
 int Transaction::askUid() const { return d_askUid; }
 const Underlying& Transaction::underlying() const { return d_underlying; }
 double Transaction::price() const { return d_price; }
-double Transaction::qnty() const { return d_qnty; }
+int Transaction::qnty() const { return d_qnty; }
 const TimePoint& Transaction::timeExecuted() const { return d_timeExecuted; }
 
 std::ostream& operator<<(std::ostream& os, const Transaction& transaction)
