+++ src/common/transaction.h
@@ -25,18 +25,18 @@ struct Transaction
     int askUid() const;
     const Underlying& underlying() const;
     double price() const;
-    double qnty() const;
+    int qnty() const;
     const TimePoint& timeExecuted() const;
 
    private:
-    Transaction(OrderPtr bid, OrderPtr ask, double price, double qnty);
+    Transaction(OrderPtr bid, OrderPtr ask, double price, int qnty);
 
     std::string d_uid;
     int d_bidUid;
     int d_askUid;
     Underlying d_underlying;
     double d_price;
-    double d_qnty;
+    int d_qnty;
     TimePoint d_timeExecuted;
 };
